requirements-dev.txt
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1592528583041/work
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1598956214003/work
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
python-decouple==3.3
pytz==2020.1
razorpay==1.2.0
requests==2.24.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.10
wincertstore==0.2
